actions change admin routes for testimonials actions change admin routes for general routes actions change admin routes for room routes list my locations cookies popup and alert in update testimonial alert

+++ src/Modules/AdminPanel/Modules/Testimonials/Store/actions.js
@@ -12,16 +12,16 @@ export const ACTION_GET_TESTIMONIAL_BY_ID = async ({ commit }, id) => {
 };
 
 export const ACTION_POST_TESTIMONIAL = async ({ commit }, testimonial) => {
-  const { data } = await EspacioAPI.post(`/testimonial/`, testimonial);
+  const { data } = await EspacioAPI.post(`/admin/testimonial/`, testimonial);
   commit("MUTATION_POST_TESTIMONIAL", data);
   return data
 };
 
 export const ACTION_PUT_TESTIMONIAL = async ({ commit }, {id,testimonial}) => {
-  const { data } = await EspacioAPI.put(`/testimonial/${id}`,testimonial);
+  const { data } = await EspacioAPI.put(`/admin/testimonial/${id}`,testimonial);
   commit("MUTATION_PUT_TESTIMONIAL", data);
 };
 export const ACTION_DELETE_TESTIMONIAL = async ({ commit }, id) => {
-  await EspacioAPI.delete(`/testimonial/${id}`);
+  await EspacioAPI.delete(`/admin/testimonial/${id}`);
   commit("MUTATION_DELETE_TESTIMONIAL", id);
 };
